Updated example program to include SimplePriorityQueue usage
