+++ Priority Queue Example/Priority Queue Example.csproj
@@ -41,6 +41,8 @@
     <Reference Include="System.Xml" />
   </ItemGroup>
   <ItemGroup>
+    <Compile Include="SimplePriorityQueueExample.cs" />
+    <Compile Include="FastPriorityQueueExample.cs" />
     <Compile Include="Program.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
   </ItemGroup>
